requirements.txt
fastapi
uvicorn
python_dotenv
psycopg2
sqlalchemy
pytest